[BUGFIX] Fix PHP fatal error in be.tableList view helper
The be.tableList view helper uses the class localRecordList
which relies on a available $GLOBALS['SOBE']->doc object.
Since https://review.typo3.org/19000/ the doc instance
in SOBE does not get instantiated in any case, which
results in a PHP fatal error using this view helper.

For 6.x this was fixed with https://review.typo3.org/15007/

Fixes: #54748
Releases: 4.7, 4.5
Change-Id: I4ac4a1718ec6c8e0a02723802bf73dbabd1648ab
Reviewed-on: https://review.typo3.org/26890
Reviewed-by: Markus Klein
Tested-by: Markus Klein