[BUGFIX] Fixes selected element label that is added by autocomplete suggest wizard.
In flexforms, the AJAX listener that returns the options that show up in the suggest wizard
returns a set of list tags. The ID on these list tags is based on the ID of the field that
will contain the selected element. With flexforms, this ID is long and full of brackets and
other characters that don't belong in HTML IDs. This revision changes how those IDs are
generated so that alphanumeric characters are removed.