[B] Work around concurrent image processing issues
The problem is that, sometimes, when multiple requests/processes are trying
to process the same image at the same time, the image processing can fail, resulting
in a blank 'identifier' field landing in the sys_file_processedfile table. In
turn, this means that the image will not ever again be processed (at those particular
dimensions) on the frontend. So you get full-sized images every time.

This isn't ideal, because if an image actually can't be processed for some reason,
TYPO3 will still try to process it every time the associated page or context is
generated. However, in most cases this is acceptable vs. having huge unprocessed images
showing up on the frontend of the site.