[S] Apply patch for TYPO3-CORE-SA-2016-013
Extbase request handling fails to implement a proper access check for requested controller/ action combinations, which makes it possible for an attacker to execute arbitrary Extbase actions by crafting a special request. To successfully exploit this vulnerability, an attacker must have access to at least one Extbase plugin or module action in a TYPO3 installation. The missing access check inevitably leads to information disclosure or remote code execution, depending on the action that an attacker is able to execute.

Please see the TYPO3 bulletin for more details:
https://typo3.org/teams/security/security-bulletins/typo3-core/typo3-core-sa-2016-013/